add status to response and create simple front with vue 3

+++ gateway/domain/customer/customer.controller.js
@@ -13,16 +13,19 @@ class CustomerController {
     const { client, request } = await findCustomer(aggregateData);
     client.findCustomer(request, (error, response) => {
       if (!error) return res.json(customerListGetter(response));
-      else return res.json(error);
+      else return res.status(400).json(error);
     });
   }
 
   async findOne(req, res) {
     const aggregateData = req.body;
+
+    if (!Object.keys(aggregateData).length) return res.status(400).json();
+
     const { client, request } = await findOneCustomer(aggregateData);
     client.findOneCustomer(request, (error, response) => {
       if (!error) return res.json(customerGetter(response));
-      else return res.json(error);
+      else return res.status(404).json(error);
     });
   }
 
@@ -32,19 +35,21 @@ class CustomerController {
     /* ---------------------------- Number Validation --------------------------- */
     const isValidPhoneNumber = phoneValidator(req.body.phoneNumber);
     if (!isValidPhoneNumber)
-      return res.json({ errors: "Phone number is not a valid number" });
+      return res
+        .status(400)
+        .json({ errors: "Phone number is not a valid number" });
     /* ---------------------------- Schema Validation --------------------------- */
     const { valid, errors } = validator(
       schema.customer.create,
       createCustomerData
     );
-    if (!valid) return res.json(errors);
+    if (!valid) return res.status(400).json(errors);
     /* -------------------------------------------------------------------------- */
 
     const { client, request } = await createCustomer(createCustomerData);
     client.createCustomer(request, (error, response) => {
       if (!error) return res.json(customerGetter(response));
-      else return res.json(error);
+      else return res.status(400).json(error);
     });
   }
 
@@ -64,16 +69,19 @@ class CustomerController {
     const { client, request } = await updateCustomer(aggregateData, updateData);
     client.updateCustomer(request, (error, response) => {
       if (!error) return res.json({ result: response.getResult() });
-      else return res.json(error);
+      else return res.status(404).json(error);
     });
   }
 
   async delete(req, res) {
     const aggregateData = req.body;
+
+    if (!Object.keys(aggregateData).length) return res.status(400).json();
+
     const { client, request } = await deleteCustomer(aggregateData);
     client.deleteCustomer(request, (error, response) => {
       if (!error) return res.json({ result: response.getResult() });
-      else return res.json(error);
+      else return res.status(404).json(error);
     });
   }
 }
